Version 1.6, Original GIT commit
1.6

Fixes a defect handling date/time values when adding/updating
1.5

Fixes a defect with defaulting select/list fields
Removes reference to People Bridge
Adds handlers available on Community but necessary to the upload process to the provided install package and instructions
Adds the ability to export and import/upload DMC Dataset data through the DMC interface.
Updates the email handler on the upload to use SMTP, not Remedy email.